semana-05/dia03/requirements.txt
django==3.2
djangorestframework
django-cors-headers
mysqlclient
Pillow